Working comparison chart.

+++ build/es/Chip/Chip.js
@@ -1,5 +1,5 @@
-import _extends from 'babel-runtime/helpers/extends';
-import _objectWithoutProperties from 'babel-runtime/helpers/objectWithoutProperties';
+import _extends from "@babel/runtime/helpers/extends";
+import _objectWithoutProperties from "@babel/runtime/helpers/objectWithoutProperties";
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
@@ -13,7 +13,6 @@ export const styles = theme => {
   const height = 32;
   const backgroundColor = theme.palette.type === 'light' ? theme.palette.grey[300] : theme.palette.grey[700];
   const deleteIconColor = fade(theme.palette.text.primary, 0.26);
-
   return {
     root: {
       fontFamily: theme.typography.fontFamily,
@@ -26,12 +25,15 @@ export const styles = theme => {
       backgroundColor,
       borderRadius: height / 2,
       whiteSpace: 'nowrap',
-      transition: theme.transitions.create(),
+      transition: theme.transitions.create(['background-color', 'box-shadow']),
       // label will inherit this from root, then `clickable` class overrides this for both
       cursor: 'default',
-      outline: 'none', // No outline on focused element in Chrome (as triggered by tabIndex prop)
-      border: 'none', // Remove `button` border
+      // We disable the focus ring for mouse, touch and keyboard users.
+      outline: 'none',
+      border: 'none',
+      // Remove `button` border
       padding: 0 // Remove `button` padding
+
     },
     clickable: {
       // Remove grey highlight
@@ -83,47 +85,70 @@ export const styles = theme => {
     }
   };
 };
-
 /**
  * Chips represent complex entities in small blocks, such as a contact.
  */
+
 class Chip extends React.Component {
   constructor(...args) {
     var _temp;
 
-    return _temp = super(...args), this.chipRef = null, this.handleDeleteIconClick = event => {
-      // Stop the event from bubbling up to the `Chip`
-      event.stopPropagation();
-      const { onDelete } = this.props;
-      if (onDelete) {
-        onDelete(event);
-      }
-    }, this.handleKeyDown = event => {
-      // Ignore events from children of `Chip`.
-      if (event.currentTarget !== event.target) {
-        return;
-      }
+    return _temp = super(...args), Object.defineProperty(this, "chipRef", {
+      configurable: true,
+      enumerable: true,
+      writable: true,
+      value: null
+    }), Object.defineProperty(this, "handleDeleteIconClick", {
+      configurable: true,
+      enumerable: true,
+      writable: true,
+      value: event => {
+        // Stop the event from bubbling up to the `Chip`
+        event.stopPropagation();
+        const {
+          onDelete
+        } = this.props;
 
-      const { onClick, onDelete, onKeyDown } = this.props;
-      const key = keycode(event);
-
-      if (onClick && (key === 'space' || key === 'enter')) {
-        event.preventDefault();
-        onClick(event);
-      } else if (onDelete && key === 'backspace') {
-        event.preventDefault();
-        onDelete(event);
-      } else if (key === 'esc') {
-        event.preventDefault();
-        if (this.chipRef) {
-          this.chipRef.blur();
+        if (onDelete) {
+          onDelete(event);
         }
       }
+    }), Object.defineProperty(this, "handleKeyDown", {
+      configurable: true,
+      enumerable: true,
+      writable: true,
+      value: event => {
+        // Ignore events from children of `Chip`.
+        if (event.currentTarget !== event.target) {
+          return;
+        }
 
-      if (onKeyDown) {
-        onKeyDown(event);
+        const {
+          onClick,
+          onDelete,
+          onKeyDown
+        } = this.props;
+        const key = keycode(event);
+
+        if (onClick && (key === 'space' || key === 'enter')) {
+          event.preventDefault();
+          onClick(event);
+        } else if (onDelete && key === 'backspace') {
+          event.preventDefault();
+          onDelete(event);
+        } else if (key === 'esc') {
+          event.preventDefault();
+
+          if (this.chipRef) {
+            this.chipRef.blur();
+          }
+        }
+
+        if (onKeyDown) {
+          onKeyDown(event);
+        }
       }
-    }, _temp;
+    }), _temp;
   }
 
   render() {
@@ -140,19 +165,27 @@ class Chip extends React.Component {
       onKeyDown,
       tabIndex: tabIndexProp
     } = _props,
-          other = _objectWithoutProperties(_props, ['avatar', 'classes', 'className', 'component', 'deleteIcon', 'label', 'onClick', 'onDelete', 'onKeyDown', 'tabIndex']);
+          other = _objectWithoutProperties(_props, ["avatar", "classes", "className", "component", "deleteIcon", "label", "onClick", "onDelete", "onKeyDown", "tabIndex"]);
 
-    const className = classNames(classes.root, { [classes.clickable]: onClick }, { [classes.deletable]: onDelete }, classNameProp);
-
+    const className = classNames(classes.root, {
+      [classes.clickable]: onClick
+    }, {
+      [classes.deletable]: onDelete
+    }, classNameProp);
     let deleteIcon = null;
+
     if (onDelete) {
       deleteIcon = deleteIconProp && React.isValidElement(deleteIconProp) ? React.cloneElement(deleteIconProp, {
         className: classNames(deleteIconProp.props.className, classes.deleteIcon),
         onClick: this.handleDeleteIconClick
-      }) : React.createElement(CancelIcon, { className: classes.deleteIcon, onClick: this.handleDeleteIconClick });
+      }) : React.createElement(CancelIcon, {
+        className: classes.deleteIcon,
+        onClick: this.handleDeleteIconClick
+      });
     }
 
     let avatar = null;
+
     if (avatarProp && React.isValidElement(avatarProp)) {
       avatar = React.cloneElement(avatarProp, {
         className: classNames(classes.avatar, avatarProp.props.className),
@@ -166,27 +199,20 @@ class Chip extends React.Component {
       tabIndex = onClick || onDelete ? 0 : -1;
     }
 
-    return React.createElement(
-      Component,
-      _extends({
-        role: 'button',
-        className: className,
-        tabIndex: tabIndex,
-        onClick: onClick,
-        onKeyDown: this.handleKeyDown,
-        ref: node => {
-          this.chipRef = node;
-        }
-      }, other),
-      avatar,
-      React.createElement(
-        'span',
-        { className: classes.label },
-        label
-      ),
-      deleteIcon
-    );
+    return React.createElement(Component, _extends({
+      role: "button",
+      className: className,
+      tabIndex: tabIndex,
+      onClick: onClick,
+      onKeyDown: this.handleKeyDown,
+      ref: node => {
+        this.chipRef = node;
+      }
+    }, other), avatar, React.createElement("span", {
+      className: classes.label
+    }, label), deleteIcon);
   }
+
 }
 
 Chip.propTypes = process.env.NODE_ENV !== "production" ? {
@@ -194,48 +220,57 @@ Chip.propTypes = process.env.NODE_ENV !== "production" ? {
    * Avatar element.
    */
   avatar: PropTypes.element,
+
   /**
    * Useful to extend the style applied to components.
    */
   classes: PropTypes.object.isRequired,
+
   /**
    * @ignore
    */
   className: PropTypes.string,
+
   /**
    * The component used for the root node.
    * Either a string to use a DOM element or a component.
    */
   component: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+
   /**
    * Override the default delete icon element. Shown only if `onDelete` is set.
    */
   deleteIcon: PropTypes.element,
+
   /**
    * The content of the label.
    */
   label: PropTypes.node,
+
   /**
    * @ignore
    */
   onClick: PropTypes.func,
+
   /**
    * Callback function fired when the delete icon is clicked.
    * If set, the delete icon will be shown.
    */
   onDelete: PropTypes.func,
+
   /**
    * @ignore
    */
   onKeyDown: PropTypes.func,
+
   /**
    * @ignore
    */
   tabIndex: PropTypes.oneOfType([PropTypes.number, PropTypes.string])
 } : {};
-
 Chip.defaultProps = {
   component: 'div'
 };
-
-export default withStyles(styles, { name: 'MuiChip' })(Chip);
+export default withStyles(styles, {
+  name: 'MuiChip'
+})(Chip);
